requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
APScheduler==3.11.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.38.2
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
cattrs==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
contourpy==1.3.2
curl_cffi==0.11.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==5.2.5
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.20.3
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.58.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2026.1.0
fundamentus==0.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.4
google-cloud-vision==3.12.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.1
grpcio-status==1.71.0
gspread==6.2.1
gunicorn==25.3.0
httplib2==0.22.0
idna==3.10
ImageIO==2.37.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
msrest==0.7.1
multidict==6.4.3
multitasking==0.0.11
mysql-connector-python==9.3.0
narwhals==1.38.0
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.3.1
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
playwright==1.58.0
plotly==6.1.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.3.0
pyparsing==3.2.3
pypiwin32==223
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.3
pyzmq==26.4.0
redis==6.4.0
referencing==0.36.2
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
scikit-image==0.26.0
scipy==1.17.0
setuptools==80.10.2
shapely==2.1.2
six==1.17.0
smmap==5.0.2
soupsieve==2.7
spotipy==2.25.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.45.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tifffile==2026.1.28
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.6.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.6
yarl==1.20.0
yfinance==0.2.61
